call specialized method instance when encountering unspecialized sparams
In some instances, the preferred compilation signature will require
sparams to be provided at runtime. When we build the cache around these,
we need to make sure the method instance we are calling has those values
computed for the current signature, and not use the widened signature.
But we can still compile for the widened signature, we just need to make
sure we create a cache entry for every narrower call signature.

Fix #47476