RFC: Allow Module as type parameters
The intended use case for this is generated functions that want to
generate some reference to a module-specific generic function. The
current solution is to duplicate the generated function into every
module (probably using a package-provided macro) or to have some
sort of registry system in the package providing the generated
function. Both of these seem a bit ugly and I don't think there's
any particularly good reason not to allow Modules to be type
parameters. Admittedly, modules are not part of the scope
contemplated by #33387 as they are mutable, but I think the mental
model of modules is that they're immutable references to a namespace
and what's actually mutable is the namespace itself (i.e. people
wouldn't expect two modules that happen to have the same content
be `==`). This makes me think it still fits the mental model.